tests/wpt/web-platform-tests/trusted-types/input-src.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.sub.js"></script>

<body>
<script>
  test(t => {
    var url = TrustedURL.create(URLS.safe);

    var d = document.createElement('input');
    d.src = url;
    assert_equals("" + d.src, URLS.safe);
  }, "src = TrustedURL.create().");

  test(t => {
    var url = TrustedURL.unsafelyCreate(URLS.safe);

    var d = document.createElement('input');
    d.src = url;
    assert_equals("" + d.src, URLS.safe);
  }, "src = TrustedURL.unsafelyCreate().");
</script>
